Tweak threading. Add imgui club memory editor.
Use BeebThread::m_mutex to control all access to the BeebThread's
BBCMicro object. Eliminate the need for the disc mutex, and simplify
general access to the BBCMicro from the main thread for the
debugger. The functionality can now just lock the mutex and do
whatever.

The mutex is now acquired more often, but it's OK, because there's
typically no contention, so the cost is minimal.