Fiddle with threading a bit (again).
Don't be clever about locking the mutex. 99% of the time, only the
emulator thread should be locking the mutex, so the overhead of
just locking it every time should be minimal.

(This isn't true when the debugger is running, but that can wait
until later. Not sure there's any easy fix for this anyway - feels
like it's just going to be too fiddly to write unless there's just
one convenient mutex letting the UI access with the emulator.
Perhaps unoptimized build performance is just going to have to be
a bit rubbish when using the integrated debugger.)