Fix a couple of dear imgui layout problems.
The window padding for the display panel should ideally be 0x0.
This has to be set before the display panel's call to BeginWindow -
but this call is in done inside BeginDock, which performs some dear
imgui calls first to draw the tab bar and so on. So the window
padding can't be changed before the call, because the tab bar ends
up a mess, and it can't be changed after the call, because then it's
too late.

But it's only 8x8. There are probably some other issues too. With
the test pattern active it's obvious that the left and right edges
of the emulator display are getting clipped, too.

This all probably wants revisiting at some point.