Rejig HTTP method handling.
Respond to HTTP requests only once the BeebThread message has been
serviced. (This is necessary to avoid race conditions with stuff
like reset, and paste, that can take a while to run.)

Remove the HTTP request queue. The code was annoying, the
interaction with requests that return a response body was a faff,
and with the response delay there's much less call for it.

(Will probably need to add stop/start requests, along with some
easy way of getting the emulator unstuck from the UI when
something goes wrong.)