Fix unions.
Unions that consist of a primitive value member and a bitfield bits
member need to have the value member first. The ={} initialization
syntax only initializes the first member listed; if this member is a
bitfield that doesn't use every bit, the remaining bits are left as
junk.

(Not sure why this didn't show up more often, but most such unions in b2
have bitfield members that do use every bit.)