Fix Opus 'Disk not formatted' error after BREAK.
Always read from the disc drive's track, so Read Address can function
when 1770 track register and drive head disagree. Check for mismatches
when doing a read or write.

(Seems Opus just leaves the drive head wherever it was when resetting
the 1770, and can (not unreasonably) initiate a Read Address with a
mismatch between 1770 track register and drive head position.)