When a VIA pin is an input, initially set that pin's value to 1.
Fix Spy Hunter, which writes sound chip commands that have bit 7 set
while DDRA=$7f.

(I think this is safe, as the pins are only set to 1 when ORx/DDRx is
changed. If anything is going to change the value, the value will only
be wrong for 1 cycle, and the CPU can't see it.)

Fix #48.