Fix VIA PA/PB values. Set slow databus with AND. Fix RTC update.
Model the way that devices connected to the VIAs can turn a 1 into a 0
(by routing it to ground?) but seemingly can't do the opposite (or at
least not in a way the VIA can sense it). The control lines should
probably do the same.

(BeebLink still behaves inauthentically, by just setting the port value
to whatever. Hardly worth fixing.)

Fixes VIA.PB6 test.

The RTC update was always a bit broken and I hadn't previously noticed.