Don't include ; in the expansion of LOG_DEFINE and LOG_TAGGED_DEFINE.
clang-format 5 can't deal with this sort of macro: https://stackoverflow.com/questions/53060515/clang-format-indentation-of-macros
(see #19), and the lack of ; is weird anyway.

(Can't remember why the ; was in there! Maybe it was to avoid stray ;s
when stripping the logging stuff out? Not an issue for b2, and the
macros can probably expand to a static_assert(true,"") or something if
this is actually a problem.)