Command2ify TogglePopups. Fix debugger-only commands and popups.
Always compile all commands in, even if the functionality is
unavailable, so that the keyboard shortcuts don't go missing.

For ordinary commands, it's the WasActioned code that gets stripped out;
for popups, the factory function returns nullptr. Additionally, the
BeebWindow will display a placeholder popup for popups of unsupported
type that are created. This stops dock setup settings getting lost.