Fix visual_studio_test_runner when ctest not on PATH.
There doesn't seem to be a specific variable for the ctest binary, but
it's probably in the same folder as cmake, so it can figure it out.

(I assume 4bd1939513b613c4ff0916eb6021b893a1a787a6 broke this? I'm sure
I've used this projects since then though! But, maybe not.)