Don't bother returning a success flag from DebugUI::ReadByte.
There's always a value available, and if it's from g_unmapped_reads
then it doesn't matter because that's the value the CPU would see
anyway. Better to use the return value for the byte the CPU would read.

An optional out parameter would be a better way of indicating failure,
inasmuch as the concept exists, if this becomes a problem.

Fix #400.