Tidy up etc/roms.
Remove ROMs not included in the release.

Exception: Instant MIni Office II, provided as a single 128 KB
image, made up by concatenating the commonly available 4 x 32 KB
images. This is provided as a convenience as b2 doesn't currently
(and may never) support loading a ROM from multiple files, and
having it in the repo makes it easy to include a link in the docs.

Fix #410.