Update VBlankMonitor to accommodate Windows 10 fast user switching.
After user switching, the DXGI stuff needs re-enumerating, something
seemingly only detectable with a polling-type API. So poll the
vblank monitor somewhat regularly (currently: same frequency as the
window title update) to give it a chance to detect this and refresh
itself.

macOS: TBC. If it needs polling, there's now a hook for that. If it
doesn't, it can be accommodated.

Linux: going to leave this as-is until somebody files a bug...

Fix #143. See #89.