Fix log tags being totally wrong.
They are strings in the code, so the # operator is not appropriate.
Introduced by d5a2e55369754e60f390107dfcdde573f1863778.

(One advantage of them being identifiers - if they actually were,
which they are not - would be that they inherently wouldn't contain
spaces, simplifying some aspects of the command line UI. But that's
easily checked with an assert, and because the command line UI uses
the table set up on startup by the global constructors, that assert
will catch everything interesting.)