Create new curl handle for each request.
Saw this suggested as a way of fixing https connection errors, and the
error codes weren't looking very illuminating, so why not.

Doesn't seem to have helped, but I'm a bit happier with this arrangement
as it's probably a good idea to recreate the curl handles more often
rather than less.