Move some cursor processing to the VideoULA class.
In visible bitmap regions, add the cursor to the pixel buffer, so it
moves with the scroll offset; in non-visible bitmap regions, add the
cursor to the output pixels, so it ignores the scroll offset. (See
#406, now hopefully fixed.)

In teletext mode, behave mostly as before - just tweak the code to
accommodate the changes. #182 still applies.