Add some nlohmann::json hooks for the config persistence.
RapidJSON is fine as far as it goes, but nlohmann::json has some very
convenient serialization stuff built in that it'd be nice to use. But
keep the existing stuff as-is, at least for now, as converting it all
over to nlohmann::json was proving a tedious slog. This is supposed to
be a fun project.

Add helpers for sneaking in some nlohmann::json handling of the odd
bit of data. By way of test, use this to add in persistent Video NuLA
flag, previously apparently completely forgotten about.

So far, so good enough. To be continued.

See #257.