Use IsItmeDeactivatedAfterEdit to handle the font size widget.
Still a bit ugly though! The widget text visibly updates
independently of the actual value. Might be difficult to fix
without doing it manually though? - the int edit is 3 widgets (text
box, +/- buttons) and the clamping wants to be done on every edit
for the buttons but only on return press for the text box.

See #461.