Don't start emulated BBC running until the main thread is ready.
Avoid the situation where the emulator gets stuck with a full video
data buffer, because the main thread isn't consuming the data yet.

This sort-of doesn't matter, but the startup mutex behaviour ends up
very different from the ongoing behaviour, messing up the stats. It'll
also save a bit of CPU time on startup.

Fix #292.